fix warning

+++ stepmania/src/ScreenOptionsMaster.cpp
@@ -594,9 +594,9 @@ void ScreenOptionsMaster::RefreshIcons()
 			if( iSelection >= (int)m_OptionRow[i].choices.size() )
 			{
 				/* Invalid selection.  Send debug output, to aid debugging. */
-				CString error = ssprintf("Option row with name '%s' selects item %i, but there are only %lu items:\n",
+				CString error = ssprintf("Option row with name '%s' selects item %i, but there are only %i items:\n",
 					m_OptionRow[i].name.c_str(),
-					iSelection, m_OptionRow[i].choices.size() );
+					iSelection, (int) m_OptionRow[i].choices.size() );
 
 				for( unsigned j = 0; j < m_OptionRow[i].choices.size(); ++j )
 					error += ssprintf("    %s\n", m_OptionRow[i].choices[j].c_str());
